Map filters initialization in a function
