+++ app/javascript/packs/application.js
@@ -1,3 +1,3 @@
 import "bootstrap";
 // import '../components/map';
-import '../components/select2';
+import '../components/map_filters';
