+++ app/views/harbours/index.html.erb
@@ -3,7 +3,7 @@
 <div class='container ml0 mr0'>
   <div class="row">
     <div class="col-xs-12 col-sm-3 mt-">
-      <div class="select2-menu-container">
+      <div class="select2-menu-container" id="map_filters">
 
         <!-- Select2 options -->
         <p class="filter-title"><%= t('.harbour_filter')%></p> <!-- harbours -->
